requirements.txt
anytree==2.13.0
black==26.1.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
docstring-to-markdown==0.17
fonttools==4.61.1
importlib_metadata==8.7.1
jedi==0.19.2
kiwisolver==1.4.9
matplotlib==3.10.8
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
parso==0.8.5
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.14.0
pytokens==0.4.1
six==1.17.0
typing_extensions==4.15.0
ujson==5.11.0
zipp==3.23.0
